+++ src/main/java/logisticspipes/pipes/basic/CoreUnroutedPipe.java
@@ -338,7 +338,7 @@ public abstract class CoreUnroutedPipe implements IClientState, ILPPipe, ILPCCTy
             }
 
             @Override
-            public int getIC2PowerLevel() {
+            public long getIC2PowerLevel() {
                 return 0;
             }
